oops

+++ pfo/commands/reset_output.py
@@ -1,4 +1,5 @@
 # system
+import os
 import shutil
 
 # constants
@@ -11,10 +12,11 @@ logger = get_logger()
 
 
 def reset_output():
-    logger.warning(f"{OUTPUT_DIRECTORY} directory exists, removing it...")
-
-    shutil.rmtree(OUTPUT_DIRECTORY)
-
-    logger.info(f"{OUTPUT_DIRECTORY} directory was reset")
+    if os.path.isdir(OUTPUT_DIRECTORY):
+        logger.warning(f"{OUTPUT_DIRECTORY} directory exists, removing it...")
+        shutil.rmtree(OUTPUT_DIRECTORY)
+        logger.info(f"{OUTPUT_DIRECTORY} directory was reset")
+    else:
+        logger.warning(f"{OUTPUT_DIRECTORY} does not exist, reset not necessary")
 
     return None
